BOAC-5035, 'mixins' are discouraged in Vue 3; move reusable functions to typescript
