+++ src/utils.ts
@@ -1,5 +1,6 @@
 import _ from 'lodash'
 import Vue from 'vue'
+import VueScrollTo from 'vue-scrollto'
 
 export function oxfordJoin(arr, zeroString) {
   switch((arr || []).length) {
@@ -23,4 +24,45 @@ export function putFocusNextTick(id: string, cssSelector?: string) {
       }
     }, 500)
   })
+}
+
+export function scrollTo(anchor) {
+  VueScrollTo.scrollTo(anchor, 400)
+}
+
+export function scrollToTop(delay) {
+  VueScrollTo.scrollTo('#app', (delay || 400))
+}
+
+export function sortComparator(a, b, nullFirst=true) {
+  if (_.isNil(a) || _.isNil(b)) {
+    if (nullFirst) {
+      return _.isNil(a) ? (_.isNil(b) ? 0 : -1) : 1
+    } else {
+      return _.isNil(b) ? (_.isNil(a) ? 0 : -1) : 1
+    }
+  } else if (_.isNumber(a) && _.isNumber(b)) {
+    return a < b ? -1 : a > b ? 1 : 0
+  } else {
+    const aInt = toInt(a)
+    const bInt = toInt(b)
+    if (aInt && bInt) {
+      return aInt < bInt ? -1 : aInt > bInt ? 1 : 0
+    } else {
+      return a.toString().localeCompare(b.toString(), undefined, {
+        numeric: true
+      })
+    }
+  }
+}
+
+export function stripHtmlAndTrim(html) {
+  let text = html && html.replace(/<([^>]+)>/ig,'')
+  text = text && text.replace(/&nbsp;/g, '')
+  return _.trim(text)
+}
+
+export function toInt(value, defaultValue = null) {
+  const parsed = parseInt(value, 10)
+  return Number.isInteger(parsed) ? parsed : defaultValue
 }
